# next.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inbound API v2",
    "version": "2.0.0",
    "description": "The official API documentation for Inbound - Email API Service. Manage domains, send emails, handle webhooks, and more.",
    "contact": {
      "name": "Inbound Support",
      "url": "https://inbound.new/support",
      "email": "support@inbound.new"
    },
    "license": {
      "name": "Commercial",
      "url": "https://inbound.new/terms"
    }
  },
  "servers": [
    {
      "url": "https://inbound.new/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v2",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API Key authentication. Use 'Bearer YOUR_API_KEY'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for session-based authentication"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required. Please provide a valid API key or session token."
                },
                "code": {
                  "type": "string",
                  "example": "401"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred. Please try again later."
                },
                "code": {
                  "type": "string",
                  "example": "500"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/v2/mail/thread-counts": {
      "post": {
        "operationId": "post-v2-mail-thread-counts",
        "summary": "Get email thread counts",
        "description": "Calculate conversation thread sizes for multiple emails in batch. Useful for inbox listings to show conversation counts.",
        "tags": [
          "Mail"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emailIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "nullable": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thread count results for each email",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "emailId": {
                            "type": "string"
                          },
                          "threadCount": {
                            "type": "number"
                          },
                          "hasThread": {
                            "type": "boolean",
                            "description": "true if thread count > 1"
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/v2/onboarding/demo": {
      "post": {
        "operationId": "post-v2-onboarding-demo",
        "summary": "Send onboarding demo email",
        "description": "Send a demo email to test the SDK integration during onboarding. The recipient can reply to complete the onboarding flow.",
        "tags": [
          "Onboarding"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apiKey": {
                    "type": "string",
                    "nullable": false
                  },
                  "to": {
                    "type": "string",
                    "nullable": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demo email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The ID of the sent demo email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}